requirements.txt
langgraph
langchain
google-generativeai
pandas
pypdf
pytest
python-dotenv
setuptools